requirements.txt
Flask==2.2.2
slack_sdk==3.21.1
python-dotenv==1.0.0
# certifi==2023.6.15
pymongo==4.4.0
openai==0.27.0
fuzzywuzzy==0.18.0
